requirements.txt
web3
eth-brownie
dotenv
pyaml_env
requests
pick
